+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased] - 2021-xx-xx
+## [v2.1.0] - 2021-07-12
 
 ### Added
 
-- Support for resend confirmation email by the user during login
+- Support for resending email verificaton when the user status is in pending confirmation
 
 #### Changed